+++ forms/src/services/errorMessagesExtractor/errorMessagesExtractor.interface.ts
@@ -1,0 +1,15 @@
+/**
+ * Object containing information about validation errors result
+ */
+export interface ValidationErrorsResult
+{
+    /**
+     * Array of error names that are present
+     */
+    errors: string[];
+
+    /**
+     * Array of error messages that should be displayed
+     */
+    errorMessages: string[];
+}